#262 refactor: Combine AuthPageHeader mobile and desktop into one component
